CourseDetails incomplete single block but implementations work
Now, I need to break it down into multiple components, to make it more manageable.
Then, I should implement availabilities so that the saving new ones will first check the diff to delete those that were removed, do nothing for those that were not changed, and  save the new ones.

There are a few messages to show the user...

And then I can move on to implementing what is seen in a Tutor's and a Student's calendar in MyDesk