Shows now other course timeslots in a course calendar
This is to avoid double booking, though in reality it is still possible by clicking on the side.

Fixed also some small things:
- Removed margin of calendar on the course popup page
- Decreased the font-size of that calendar on small screens.

For students, it is a bit more tricky to show other bookings. The bookings from the API only show the id and nothing about the start time and end time. for that, I need to get back to the availability.